add more access controls

+++ src/components/users/users_controller.js
@@ -50,6 +50,10 @@ export default class UsersController {
   };
 
   updateName = async (req, res) => {
+    if (!this.usersService.isPermitted(req, ['OWNER', 'ADMIN'])) {
+      return this.utils.fail(res, { message: 'You are not allowed to perform that action.' }, 403);
+    }
+
     const user = req.body;
 
     const objectToUpdate = {
@@ -61,6 +65,10 @@ export default class UsersController {
   };
 
   updateEmail = async (req, res) => {
+    if (!this.usersService.isPermitted(req, ['OWNER', 'ADMIN'])) {
+      return this.utils.fail(res, { message: 'You are not allowed to perform that action.' }, 403);
+    }
+
     const user = req.body;
 
     const objectToUpdate = {
@@ -71,6 +79,10 @@ export default class UsersController {
   };
 
   updatePassword = async (req, res) => {
+    if (!this.usersService.isPermitted(req, ['OWNER', 'ADMIN'])) {
+      return this.utils.fail(res, { message: 'You are not allowed to perform that action.' }, 403);
+    }
+
     const user = req.body;
 
     const saltAndHashPassword = await this.usersService.saltAndHashPassword(user.password);
@@ -82,7 +94,11 @@ export default class UsersController {
     this.usersService.updateUser(req, res, user.id, objectToUpdate);
   };
 
-  updateUser = async (req, res) => {
+  updateWholeUser = async (req, res) => {
+    if (!this.usersService.isPermitted(req, ['ADMIN'])) {
+      return this.utils.fail(res, { message: 'You are not allowed to perform that action.' }, 403);
+    }
+
     const user = req.body;
     const objectToUpdate = {
       first_name: user.firstName,
@@ -96,6 +112,24 @@ export default class UsersController {
   };
 
   deleteUser = async (req, res) => {
+    if (!this.usersService.isPermitted(req, ['OWNER', 'ADMIN'])) {
+      return this.utils.fail(res, { message: 'You are not allowed to perform that action.' }, 403);
+    }
+
+    const user = req.body;
+
+    const objectToUpdate = {
+      status: 'DELETED',
+    };
+
+    this.usersService.updateUser(req, res, user.id, objectToUpdate);
+  }
+
+  removeUserFromDatabase = async (req, res) => {
+    if (!this.isPermitted(req, ['ADMIN'])) {
+      return this.utils.fail(res, { message: 'You are not allowed to perform that action.' }, 403);
+    }
+
     try {
       const result = await this.model.destroy({
         where: {
